vm: remove CHECKSIG/VERIFY/CHECKMULTISIG opcodes
