+++ pkg/vm/vm_test.go
@@ -8,8 +8,6 @@ import (
 	"math/rand"
 	"testing"
 
-	"github.com/nspcc-dev/neo-go/pkg/crypto/hash"
-	"github.com/nspcc-dev/neo-go/pkg/crypto/keys"
 	"github.com/nspcc-dev/neo-go/pkg/io"
 	"github.com/nspcc-dev/neo-go/pkg/util"
 	"github.com/nspcc-dev/neo-go/pkg/vm/emit"
@@ -2601,292 +2599,6 @@ func TestREMOVEMap(t *testing.T) {
 	runVM(t, vm)
 	assert.Equal(t, 1, vm.estack.Len())
 	assert.Equal(t, makeStackItem(false), vm.estack.Pop().value)
-}
-
-func TestCHECKSIGNoArgs(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	vm := load(prog)
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKSIGOneArg(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.estack.PushVal(pbytes)
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKSIGNoSigLoaded(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := "NEO - An Open Network For Smart Economy"
-	sig := pk.Sign([]byte(msg))
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.estack.PushVal(sig)
-	vm.estack.PushVal(pbytes)
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKSIGBadKey(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig := pk.Sign(msg)
-	pbytes := pk.PublicKey().Bytes()[:4]
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal(sig)
-	vm.estack.PushVal(pbytes)
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKSIGWrongSig(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig := pk.Sign(msg)
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal(util.ArrayReverse(sig))
-	vm.estack.PushVal(pbytes)
-	runVM(t, vm)
-	assert.Equal(t, 1, vm.estack.Len())
-	assert.Equal(t, false, vm.estack.Pop().Bool())
-}
-
-func TestCHECKSIGGood(t *testing.T) {
-	prog := makeProgram(opcode.CHECKSIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig := pk.Sign(msg)
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal(sig)
-	vm.estack.PushVal(pbytes)
-	runVM(t, vm)
-	assert.Equal(t, 1, vm.estack.Len())
-	assert.Equal(t, true, vm.estack.Pop().Bool())
-}
-
-func TestVERIFYGood(t *testing.T) {
-	prog := makeProgram(opcode.VERIFY)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig := pk.Sign(msg)
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.estack.PushVal(msg)
-	vm.estack.PushVal(sig)
-	vm.estack.PushVal(pbytes)
-	runVM(t, vm)
-	assert.Equal(t, 1, vm.estack.Len())
-	assert.Equal(t, true, vm.estack.Pop().Bool())
-}
-
-func TestVERIFYBad(t *testing.T) {
-	prog := makeProgram(opcode.VERIFY)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig := pk.Sign(msg)
-	pbytes := pk.PublicKey().Bytes()
-	vm := load(prog)
-	vm.estack.PushVal(util.ArrayReverse(msg))
-	vm.estack.PushVal(sig)
-	vm.estack.PushVal(pbytes)
-	runVM(t, vm)
-	assert.Equal(t, 1, vm.estack.Len())
-	assert.Equal(t, false, vm.estack.Pop().Bool())
-}
-
-func TestCHECKMULTISIGNoArgs(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	vm := load(prog)
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKMULTISIGOneArg(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	pk, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	vm := load(prog)
-	pbytes := pk.PublicKey().Bytes()
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(pbytes)})
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKMULTISIGNotEnoughKeys(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	pk1, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	pk2, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig1 := pk1.Sign(msg)
-	sig2 := pk2.Sign(msg)
-	pbytes1 := pk1.PublicKey().Bytes()
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(sig1), NewByteArrayItem(sig2)})
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(pbytes1)})
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKMULTISIGNoHash(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	pk1, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	pk2, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig1 := pk1.Sign(msg)
-	sig2 := pk2.Sign(msg)
-	pbytes1 := pk1.PublicKey().Bytes()
-	pbytes2 := pk2.PublicKey().Bytes()
-	vm := load(prog)
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(sig1), NewByteArrayItem(sig2)})
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(pbytes1), NewByteArrayItem(pbytes2)})
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKMULTISIGBadKey(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	pk1, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	pk2, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig1 := pk1.Sign(msg)
-	sig2 := pk2.Sign(msg)
-	pbytes1 := pk1.PublicKey().Bytes()
-	pbytes2 := pk2.PublicKey().Bytes()[:4]
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(sig1), NewByteArrayItem(sig2)})
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(pbytes1), NewByteArrayItem(pbytes2)})
-	checkVMFailed(t, vm)
-}
-
-func TestCHECKMULTISIGBadSig(t *testing.T) {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	pk1, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	pk2, err := keys.NewPrivateKey()
-	assert.Nil(t, err)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-	sig1 := pk1.Sign(msg)
-	sig2 := pk2.Sign(msg)
-	pbytes1 := pk1.PublicKey().Bytes()
-	pbytes2 := pk2.PublicKey().Bytes()
-	vm := load(prog)
-	vm.SetCheckedHash(hash.Sha256(msg).BytesBE())
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(util.ArrayReverse(sig1)), NewByteArrayItem(sig2)})
-	vm.estack.PushVal([]StackItem{NewByteArrayItem(pbytes1), NewByteArrayItem(pbytes2)})
-	runVM(t, vm)
-	assert.Equal(t, 1, vm.estack.Len())
-	assert.Equal(t, false, vm.estack.Pop().Bool())
-}
-
-func initCHECKMULTISIG(msg []byte, n int) ([]StackItem, []StackItem, map[string]*keys.PublicKey, error) {
-	var err error
-
-	keyMap := make(map[string]*keys.PublicKey)
-	pkeys := make([]*keys.PrivateKey, n)
-	pubs := make([]StackItem, n)
-	for i := range pubs {
-		pkeys[i], err = keys.NewPrivateKey()
-		if err != nil {
-			return nil, nil, nil, err
-		}
-
-		pk := pkeys[i].PublicKey()
-		data := pk.Bytes()
-		pubs[i] = NewByteArrayItem(data)
-		keyMap[string(data)] = pk
-	}
-
-	sigs := make([]StackItem, n)
-	for i := range sigs {
-		sig := pkeys[i].Sign(msg)
-		sigs[i] = NewByteArrayItem(sig)
-	}
-
-	return pubs, sigs, keyMap, nil
-}
-
-func subSlice(arr []StackItem, indices []int) []StackItem {
-	if indices == nil {
-		return arr
-	}
-
-	result := make([]StackItem, len(indices))
-	for i, j := range indices {
-		result[i] = arr[j]
-	}
-
-	return result
-}
-
-func initCHECKMULTISIGVM(t *testing.T, n int, ik, is []int) *VM {
-	prog := makeProgram(opcode.CHECKMULTISIG)
-	v := load(prog)
-	msg := []byte("NEO - An Open Network For Smart Economy")
-
-	v.SetCheckedHash(hash.Sha256(msg).BytesBE())
-
-	pubs, sigs, _, err := initCHECKMULTISIG(msg, n)
-	require.NoError(t, err)
-
-	pubs = subSlice(pubs, ik)
-	sigs = subSlice(sigs, is)
-
-	v.estack.PushVal(sigs)
-	v.estack.PushVal(pubs)
-
-	return v
-}
-
-func testCHECKMULTISIGGood(t *testing.T, n int, is []int) {
-	v := initCHECKMULTISIGVM(t, n, nil, is)
-
-	runVM(t, v)
-	assert.Equal(t, 1, v.estack.Len())
-	assert.True(t, v.estack.Pop().Bool())
-}
-
-func TestCHECKMULTISIGGood(t *testing.T) {
-	t.Run("3_1", func(t *testing.T) { testCHECKMULTISIGGood(t, 3, []int{1}) })
-	t.Run("2_2", func(t *testing.T) { testCHECKMULTISIGGood(t, 2, []int{0, 1}) })
-	t.Run("3_3", func(t *testing.T) { testCHECKMULTISIGGood(t, 3, []int{0, 1, 2}) })
-	t.Run("3_2", func(t *testing.T) { testCHECKMULTISIGGood(t, 3, []int{0, 2}) })
-	t.Run("4_2", func(t *testing.T) { testCHECKMULTISIGGood(t, 4, []int{0, 2}) })
-	t.Run("10_7", func(t *testing.T) { testCHECKMULTISIGGood(t, 10, []int{2, 3, 4, 5, 6, 8, 9}) })
-	t.Run("12_9", func(t *testing.T) { testCHECKMULTISIGGood(t, 12, []int{0, 1, 4, 5, 6, 7, 8, 9}) })
-}
-
-func testCHECKMULTISIGBad(t *testing.T, n int, ik, is []int) {
-	v := initCHECKMULTISIGVM(t, n, ik, is)
-
-	runVM(t, v)
-	assert.Equal(t, 1, v.estack.Len())
-	assert.False(t, v.estack.Pop().Bool())
-}
-
-func TestCHECKMULTISIGBad(t *testing.T) {
-	t.Run("1_1 wrong signature", func(t *testing.T) { testCHECKMULTISIGBad(t, 2, []int{0}, []int{1}) })
-	t.Run("3_2 wrong order", func(t *testing.T) { testCHECKMULTISIGBad(t, 3, []int{0, 2}, []int{2, 0}) })
-	t.Run("3_2 duplicate sig", func(t *testing.T) { testCHECKMULTISIGBad(t, 3, nil, []int{0, 0}) })
 }
 
 func TestSWAPGood(t *testing.T) {
